Reorganise and add labels
